+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped the minimum supported Blockbook version to [commit `95ee9b5b`](https://github.com/trezor/blockbook/commit/95ee9b5b).
 - Bumped `bitcoin` to version 0.31.
 - Adopted the block and transaction version to the corresponding `bitcoin` type.
+- Added a `bdk` feature that gates functionality to use Blockbook as a data provider for
+  the popular [`bdk`](https://crates.io/crates/bdk) wallet library.
 
 ## 0.1.0